moved selection operation constraints to a separate file
